Refactor: Moved all relevant files in to questionnaireDetails directory.
Refs: BLAIS5-3651
