+++ src/components/questionnaireDetails/blaiseNodeInfo.tsx
@@ -1,6 +1,6 @@
 import { Questionnaire } from "blaise-api-node-client";
 import { Collapsible } from "blaise-design-system-react-components";
-import QuestionnaireStatus from "../../questionnaireStatus";
+import QuestionnaireStatus from "../questionnaireStatus";
 import React, { Fragment, ReactElement } from "react";
 
 interface NodeProps {
